test(file): scope the ownership tests to unix, where ownership exists
os.Getuid and os.Getgid return -1 on Windows, so tests that format them build the string "-1" --
resolveOwnership reads it as a numeric id and then rejects it, because both sides land on the -1
"leave unchanged" sentinel.

The failure is the platform having no such concept, not the code being wrong, so the two uid-dependent
tests move behind //go:build unix in helpers_unix_test.go, matching provider_unix_test.go. The rest of the
ownership tests are platform-neutral and stay.
