+++ pkg/op/provider/file/helpers_test.go
@@ -6,7 +6,6 @@ package file
 import (
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"testing"
 )
@@ -67,24 +66,6 @@ func TestResolveOwnership_RejectsInvalid(t *testing.T) {
 	}
 }
 
-func TestResolveOwnership_LooksUpNamedUser(t *testing.T) {
-
-	// Looking up the current user by name should always succeed and return the same uid as os.Getuid.
-	currentUID := os.Getuid()
-	currentName := strconv.Itoa(currentUID)
-
-	gotUID, gotGID, err := resolveOwnership(currentName, "")
-	if err != nil {
-		t.Fatalf("resolveOwnership(%q, \"\"): %v", currentName, err)
-	}
-	if gotUID != currentUID {
-		t.Errorf("uid: got %d, want %d", gotUID, currentUID)
-	}
-	if gotGID != -1 {
-		t.Errorf("gid: got %d, want -1 (group side absent)", gotGID)
-	}
-}
-
 // --- applyOwnership ---
 
 func TestApplyOwnership_BothSidesEmptyIsNoOp(t *testing.T) {
@@ -99,22 +80,6 @@ func TestApplyOwnership_BothSidesEmptyIsNoOp(t *testing.T) {
 	// Both sides empty must short-circuit to nil error without invoking any syscall.
 	if err := applyOwnership(target, "", ""); err != nil {
 		t.Errorf("applyOwnership(empty, empty): %v", err)
-	}
-}
-
-func TestApplyOwnership_CurrentUserIsNoOp(t *testing.T) {
-
-	// Setting ownership to the current uid and gid is a no-op-ish operation that needs no CAP_CHOWN.
-	tmp := t.TempDir()
-	target := filepath.Join(tmp, "test.txt")
-
-	if err := os.WriteFile(target, []byte("test"), 0o644); err != nil {
-		t.Fatalf("WriteFile: %v", err)
-	}
-
-	user, group := strconv.Itoa(os.Getuid()), strconv.Itoa(os.Getgid())
-	if err := applyOwnership(target, user, group); err != nil {
-		t.Errorf("applyOwnership(%q, %q): %v", user, group, err)
 	}
 }
